+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/navansatt/model/NavAnsattRolle.kt
@@ -27,3 +27,15 @@ data class NavAnsattRolle(
         }
     }
 }
+
+object NavAnsattRolleHelper {
+    fun hasRole(
+        roles: Set<NavAnsattRolle>,
+        requiredRole: NavAnsattRolle,
+    ): Boolean = when (requiredRole.generell) {
+        true -> roles.any { it.rolle == requiredRole.rolle }
+        false -> roles.any {
+            it.rolle == requiredRole.rolle && (it.generell || it.enheter.containsAll(requiredRole.enheter))
+        }
+    }
+}
